fix : changed data types

+++ backend/src/entities/nft_metadata.rs
@@ -12,7 +12,7 @@ pub struct Model{
     pub symbol : String,
     #[sea_orm(column_type = "Text")]
     pub uri : String,
-    pub seller_fee_basis_pioints : i32,
+    pub seller_fee_basis_pioints : u16,
     pub update_authority : String,
     pub primary_sale_happened : bool,
     pub is_mutable : bool, // tells wheather the metadata can be changed or updated
